Создал папку Models с файлами: QuizeStepViewModel, QuizeResultsViewModel, QuizeQuestion.
