+++ MovieQuiz/Models/QuizQuestion.swift
@@ -1,0 +1,15 @@
+//
+//  QuizQuestion.swift
+//  MovieQuiz
+//
+//  Created by Андрей Мерзликин on 19.06.2023.
+//
+
+import Foundation
+import UIKit
+
+struct QuizQuestion {
+  let image: String
+  let text: String
+  let correctAnswer: Bool
+}
