+++ MovieQuiz/Models/QuizResultsViewModel.swift
@@ -1,0 +1,15 @@
+//
+//  QuizResultsViewModel.swift
+//  MovieQuiz
+//
+//  Created by Андрей Мерзликин on 19.06.2023.
+//
+
+import Foundation
+import UIKit
+
+struct QuizResultsViewModel {
+  let title: String
+  let text: String
+  let buttonText: String
+}
